backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
scikit-learn
pandas
joblib
python-dotenv
python-multipart
PyJWT
